Making the move to stdint.h types. Also, HEADER_SIZE is now calculated dynamiclly, which makes working on 64bit platforms easier.

+++ shotgun/lib/heap.h
@@ -1,7 +1,7 @@
 #ifndef __RUBINIUS_HEAP__
 #define __RUBINIUS_HEAP__ 1
 
-typedef unsigned long address;
+typedef uintptr_t address;
 
 struct heap {
   int size;
